backend/requirements.txt
backports.ssl-match-hostname==3.4.0.2
binaryornot==0.4.0
blessings==1.6
boto==2.38.0
certifi==2015.9.6.2
chardet==2.3.0
click==5.1
cookiecutter==1.2.1
coverage==3.7.1
dj-database-url==0.3.0
Django==1.8.4
-e git+https://github.com/alfetopito/django-configurations.git@71b5142354aceb5461d29e5d9e1cdf30b64b657f#egg=django_configurations-master
django-crispy-forms==1.5.2
django-filter==0.11.0
django-model-utils==2.3.1
django-nose==1.4.1
django-pnm==0.1.6
django-rq==0.8.0
django-versatileimagefield==1.0.4
djangorestframework==3.2.3
djangorestframework-jwt==1.7.2
ecdsa==0.13
Fabric==1.10.1
factory-boy==2.5.2
fake-factory==0.5.2
flake8==2.4.1
funcsigs==0.4
future==0.15.2
geopy==1.11.0
gnureadline==6.3.3
ipdb==0.8
ipython==3.1.0
Jinja2==2.8
livereload==2.4.0
Markdown==2.6.2
MarkupSafe==0.23
mccabe==0.3.1
mkdocs==0.14.0
mock==1.3.0
nose==1.3.7
nose-progressive==1.5.1
paramiko==1.15.4
pbr==1.8.1
pep8==1.5.7
Pillow==3.0.0
psycopg2==2.6
pycrypto==2.6.1
pyflakes==0.8.1
PyJWT==1.4.0
pytz==2015.4
PyYAML==3.11
redis==2.10.5
requests==2.8.1
rq==0.5.6
ruamel.base==1.0.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.10.12
six==1.10.0
tornado==4.2.1
whichcraft==0.1.1
whitenoise==2.0.3
